# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "https://github.com/junun/admin-go",
        "contact": {
            "name": "Junun",
            "url": "https://github.com/junun/admin-go",
            "email": "junun717@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
...
        "/admin/roles": {
            "get": {
                "description": "角色列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "角色列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\": 500, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "新增角色",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "新增角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "角色信息",
                        "name": "Data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.RoleResource"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\": 500, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/roles/{id}": {
            "put": {
                "description": "修改角色",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "修改角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "角色信息",
                        "name": "Data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.RoleResource"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\": 500, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除角色",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "删除角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "角色id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\": 500, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/roles/{id}/permissions": {
            "get": {
                "description": "角色权限详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "角色权限详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\": 500, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "添加/修改角色权限",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "添加/修改角色权限",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "角色权限信息",
                        "name": "Data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.RolePermResource"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\": 500, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.RolePermResource": {
            "type": "object",
            "properties": {
                "codes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "admin.RoleResource": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}